docs: operator based installation for the VPP dataplane (#5362)
* docs: update operator charts

Bring in the Installation and APIServer CRD definitions
from the current operator master.

* Update VPP dataplane instalatino docs to use the operator

* Remove APIServer from tigera-operator.yaml

The presence of that object causes the yaml to fail when it is first
applied to a cluster, because there seems to be a delay between the
creation of a CRD and the time when an object of that type can be first
created.
Since the APIServer is present in the custom-resources.yaml manifest
which is used later in the installation process, removing it from this
yaml shouldn't break anything.

* Update VPP EKS-DPDK script options

* Update hugepage requirements for VPP

* Address review comments for VPP docs