Add enPxxx support to BPFDataIfacePattern
For the Linux network device naming rules talked in [0],[1]:
prefix [Pdomain] pbus sdevice [ffunction] [nphys_port_name | ddev_port]

The domain is omitted if it's 0 and the function is only present if
it's a multi-function device. But for some network devices and/or Linux
system, the domain is non-zero, so the pattern would be "enPxxxxxx".

On the other side, from the real case showed in our server,
the ethernet interface name like "enP1p1s0f1" is valid as
a normal data interface.

Now we lack this pattern in BPFDataIfacePattern, so for the bpf's ifstate map,
this interface would be skipped as:

felix/bpf_ep_mgr.go 516: Ignoring interface that's neither data nor workload.
update=&intdataplane.ifaceUpdate{Name:"enP1p1s0f1", State:"up", Index:4}

This lack would affect the normal work of eBPF communication since it would lack
in the eBPF ifstate map which would be used in cali_iface_lookup_elem(&ifindex).

This PR is also part of the fix for the issue of:
https://github.com/projectcalico/calico/issues/6392

[0] https://utcc.utoronto.ca/~cks/space/blog/linux/UdevNetworkDeviceNaming
[1] https://www.freedesktop.org/software/systemd/man/systemd.net-naming-scheme.html

Signed-off-by: trevor tao <trevor.tao@arm.com>