Add static routes for BGP sessions for AKS Route Servers (#7149)
EGW in AKS relies on BGP sessions with Azure Route Servers. However, it needs a static route to prevent route flapping and as such session flapping. This PR adds a fields to BGPPeer spec called reachableBy for this purpose. When it is set, an exact static route (/32 for IPv4 and /128 for IPv6) is added to Bird's config file.