Add installation reference for Image Assurance
This commit copies what was done for generating the installation reference for the tigera operator and applies it to the image assurance operator.

I couldn't push to the personal for of the api-gen repo that's being used, so for this I just created my own fork that has the updated go.mod / go.sum.