Add documentation for enabling / configuring the runtime view
This PR separates out the details of the runtime view into it's own document, and does a more comprehensive overview of what it does, how you enable it, and how it needs to be configured in special cases.

It outlines pre requisites around requiring k8s clusters that use CRI, and highlights what components are installed and what privileges they need.