Update cluster-scanner.mdx
Updating the title of this topic to be more explicit that the images you're scanning are running in the cluster.  Next to the CLI scanner topic, someone scanning the topic as originally written might think that this is a version of the scanner that runs in a k8s cluster but could be used to scan anything.  Or that this could be used in conjunction with pipelines that run in a k8s cluster