Doc for DGA detection
We are recreating the DGA detector within the Runtime Security architecture.

The user will see security events for DNS lookups for algorithm-generated
domain names, if such lookups are occurring in their cluster.  In the previous
'Anomaly Detection' architecture this would have been subject to the existence
of a GlobalAlert resource with type: AnomalyDetection and detector: dga, but
now it happens unconditionally.

This change also simplifies and improves the table of MITRE tactics on the
Security Events page.  The 'Attack techniques' column is renamed 'Purpose'
because it really describes the general purpose of the tactic and not specific
techniques.  The 'Target' column is removed because it was inconsistent in its
content - sometimes an actual target, and sometimes a more general concept -
and in all cases was simply duplicating what was already in the following
column.