Update IA API reference to latest release 1.17.7 and fix api generation
IA API reference docs are quite out of date, mainly because generating
the API docs has been broken.

I've changed the generation logic to not use custom forks anymore, and
to use a much newer version of github.com/ahmetb/gen-crd-api-reference-docs
The latest one is compatible with the latest go versions, and this new
process (well the process outlined in the repo itself) doesn't require
modifying the go.mod file.

We now check out the gen repo, build it, check out the ia repo, then run
the gen tool in the operator repo.

This commit also omits internal packages from IA using the config.json
in this repo. The template is also stored in this repo in the
crd-gen-template, so any modifications to the generated code structure
can be handled there.