Apply operator-crds.yaml before creating custom resources
Add command to create/apply operator-crds.yaml before
custom-resources.yaml (installation, etc) to ensure CRDs
are present.

With the recent changes to the operator to manage CRDs, they are
no longer included in the tigera-operator.yaml manifest, so there
might be timing issues where the installation and other custom
resources are attempted to be created before their definitions are
present in the cluster.